+++ src/main/java/com/buuz135/industrial/tile/block/MobDuplicatorBlock.java
@@ -20,6 +20,7 @@ public class MobDuplicatorBlock extends CustomOrientedBlock<MobDuplicatorTile> {
 
     public int essenceNeeded;
     public List<String> blacklistedEntities;
+    public boolean enableExactCopy;
 
     public MobDuplicatorBlock() {
         super("mob_duplicator", MobDuplicatorTile.class, Material.ROCK, 5000, 80);
@@ -40,6 +41,7 @@ public class MobDuplicatorBlock extends CustomOrientedBlock<MobDuplicatorTile> {
         super.getMachineConfig();
         essenceNeeded = CustomConfiguration.config.getInt("essenceNeeded", "machines" + Configuration.CATEGORY_SPLITTER + this.getRegistryName().getResourcePath().toString(), 12, 1, Integer.MAX_VALUE, "Essence needed based on mob's health (mobHealth*essenceNeeded)");
         blacklistedEntities = Arrays.asList(CustomConfiguration.config.getStringList("blacklistedEntities", "machines" + Configuration.CATEGORY_SPLITTER + this.getRegistryName().getResourcePath().toString(), new String[]{}, "A list of blacklisted entities like minecraft:creeper"));
+        enableExactCopy = CustomConfiguration.config.getBoolean("enableExactCopy","machines" + Configuration.CATEGORY_SPLITTER + this.getRegistryName().getResourcePath().toString(), false, "Set to true to enable exact copy in the Mob Duplicator.");
     }
 
     @Override